nci-agency/anet#2959: Add generate core models functionality
