--- anet-tools/src/examples/models.py
@@ -1,263 +1,0 @@
-import copy
-import datetime
-import uuid
-
-from sqlalchemy import Boolean, Column, DateTime, Float, ForeignKey, Integer, LargeBinary, String, Text, text
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import relationship
-from sqlalchemy_mixins import ActiveRecordMixin
-
-from src.core.base_methods import base_methods
-
-
-Base = declarative_base()
-metadata = Base.metadata
-
-
-class BaseModel(Base, ActiveRecordMixin):
-    __abstract__ = True
-
-
-class anet_logic_mixin(BaseModel):
-    __abstract__ = True
-
-    def insert_entity(self, createdAt, update_rules):
-        """Insert and flush a new record
-        """
-        self.createdAt = createdAt
-        self.updatedAt = createdAt
-        if self.__tablename__ == "people":
-            utc_now = datetime.datetime.now()
-            PeoplePositions.create(createdAt=utc_now, person=self)
-        else:
-            BaseModel.session.add(self)
-            BaseModel.session.flush()
-        if self.__tablename__ == "reports":
-            # print(f"self.people {self.people}")
-            for rp in self.people:
-                # print(f"rp {rp}")
-                # print(f"rp.person {rp.person}")
-                prs = rp.person
-                #if not base_methods.is_entity_update(prs, update_rules):
-                    # print(f"rp.person {vars(rp.person)}")
-                PeoplePositions.create(createdAt=createdAt, person=rp.person)
-                BaseModel.session.flush()
-
-    def update_entity(self, utc_now, update_rules):
-        """Update and flush an existing record
-        """
-        obj = type(self).find(self.uuid)
-
-        if self.__tablename__ == "reports" and len(self.people) != 0:
-            #rp_list = copy.deepcopy(self.people)
-            for rp in self.people:
-                prs = rp.person
-                prs.set_session(self.session)
-                if base_methods.is_entity_update(prs, update_rules):
-                    delattr(rp.person, "reports")
-                    prs.updatedAt = utc_now
-                    prs.update_entity(utc_now, update_rules)
-                else:
-                    delattr(rp, "report")
-                    rp.person.createdAt = utc_now
-                    rp.person.updatedAt = utc_now
-                    obj.people.append(rp)
-                    BaseModel.session.flush()
-                    PeoplePositions.create(createdAt=utc_now, person=rp.person)
-                    BaseModel.session.flush()
-            delattr(self, "people")
-
-        self.updatedAt = utc_now
-        for attr, value in self.__dict__.items():
-            if attr != "_sa_instance_state":
-                setattr(obj, attr, value)
-        BaseModel.session.flush()
-
-    def insert_update_nested_entity(self, utc_now, update_rules):
-        self_c = copy.deepcopy(self)
-        
-        if base_methods.has_entity_relation(self, "person"):
-            base_methods.position_relation_process(
-                self, "person", self_c, update_rules, PeoplePositions, utc_now)
-
-        if base_methods.has_entity_relation(self, "location"):
-            base_methods.position_relation_process(
-                self, "location", self_c, update_rules, PeoplePositions, utc_now)
-
-        if base_methods.has_entity_relation(self, "organization"):
-            base_methods.position_relation_process(
-                self, "organization", self_c, update_rules, PeoplePositions, utc_now)
-
-        if base_methods.is_entity_update(self, update_rules):
-            if base_methods.has_entity_relation(self, "person"):
-                base_methods.remove_positions_association_with_person(
-                    self, PeoplePositions, utc_now)
-            self_c.update_entity(utc_now)
-        else:
-            self_c.insert_entity(utc_now, update_rules)
-
-        if base_methods.has_entity_relation(self, "person"):
-            base_methods.add_new_association(self, PeoplePositions, utc_now)
-
-        BaseModel.session.flush()
-
-    @classmethod
-    def commit(cls):
-        cls.session.commit()
-
-
-class PeoplePositions(anet_logic_mixin):
-    __tablename__ = "peoplePositions"
-    createdAt = Column('createdAt', DateTime)
-    personUuid = Column('personUuid', ForeignKey('people.uuid'), index=True)
-    positionUuid = Column('positionUuid', ForeignKey(
-        'positions.uuid'), index=True)
-    endedAt = Column('endedAt', DateTime)
-    __mapper_args__ = {
-        "primary_key": [createdAt, personUuid, positionUuid]
-    }
-
-    person = relationship("Person", back_populates="positions")
-    position = relationship("Position", back_populates="people")
-
-
-class Position(anet_logic_mixin):
-    __tablename__ = 'positions'
-
-    code = Column(String(100), unique=True)
-    name = Column(String(512), nullable=False, index=True)
-    createdAt = Column(DateTime, index=True)
-    updatedAt = Column(DateTime)
-    type = Column(Integer, nullable=False)
-    status = Column(Integer, nullable=False, server_default=text("0"))
-    uuid = Column(String(36), primary_key=True)
-    currentPersonUuid = Column(ForeignKey('people.uuid'), unique=True)
-    locationUuid = Column(ForeignKey('locations.uuid'), index=True)
-    organizationUuid = Column(ForeignKey('organizations.uuid'), index=True)
-
-    person = relationship("Person")
-    location = relationship('Location')
-    organization = relationship('Organization')
-    people = relationship("PeoplePositions", back_populates="position")
-
-
-class Person(anet_logic_mixin):
-    __tablename__ = 'people'
-
-    name = Column(String(255), index=True)
-    status = Column(Integer)
-    emailAddress = Column(String(255))
-    phoneNumber = Column(String(100))
-    rank = Column(String(255), index=True)
-    biography = Column(Text)
-    createdAt = Column(DateTime, index=True)
-    updatedAt = Column(DateTime)
-    role = Column(Integer, nullable=False)
-    pendingVerification = Column(Boolean, server_default=text("false"))
-    domainUsername = Column(String(500))
-    country = Column(String(500))
-    gender = Column(String(64))
-    endOfTourDate = Column(DateTime)
-    uuid = Column(String(36), primary_key=True)
-    code = Column(String(100))
-    customFields = Column(Text)
-    avatar = Column(LargeBinary)
-
-    positions = relationship("PeoplePositions", back_populates="person")
-    reports = relationship("ReportPeople", back_populates="person")
-
-
-class Location(anet_logic_mixin):
-    __tablename__ = 'locations'
-
-    name = Column(String(500), nullable=False, index=True)
-    lat = Column(Float(53))
-    lng = Column(Float(53))
-    createdAt = Column(DateTime, index=True)
-    updatedAt = Column(DateTime)
-    status = Column(Integer, nullable=False, server_default=text("0"))
-    uuid = Column(String(36), primary_key=True)
-    customFields = Column(Text)
-
-
-class Organization(anet_logic_mixin):
-    __tablename__ = 'organizations'
-
-    shortName = Column(String(255), index=True)
-    createdAt = Column(DateTime, index=True)
-    updatedAt = Column(DateTime)
-    type = Column(Integer, nullable=False, index=True)
-    longName = Column(Text, index=True)
-    identificationCode = Column(String(100), unique=True)
-    status = Column(Integer, nullable=False, server_default=text("0"))
-    uuid = Column(String(36), primary_key=True)
-    parentOrgUuid = Column(ForeignKey('organizations.uuid'), index=True)
-    customFields = Column(Text)
-
-    parent = relationship('Organization', remote_side=[uuid])
-
-
-class ApprovalStep(Base):
-    __tablename__ = 'approvalSteps'
-
-    name = Column(String(255))
-    uuid = Column(String(36), primary_key=True)
-    relatedObjectUuid = Column(String(36), nullable=False, index=True)
-    nextStepUuid = Column(ForeignKey('approvalSteps.uuid'), index=True)
-    type = Column(Integer, nullable=False)
-    restrictedApproval = Column(Boolean, server_default=text("false"))
-
-    parent = relationship('ApprovalStep', remote_side=[uuid])
-
-
-class Report(anet_logic_mixin):
-    __tablename__ = 'reports'
-
-    createdAt = Column(DateTime, index=True)
-    updatedAt = Column(DateTime, index=True)
-    intent = Column(Text)
-    exsum = Column(Text)
-    text = Column(Text)
-    nextSteps = Column(Text)
-    state = Column(Integer, nullable=False)
-    engagementDate = Column(DateTime, index=True)
-    atmosphere = Column(Integer)
-    atmosphereDetails = Column(Text)
-    keyOutcomes = Column(Text)
-    cancelledReason = Column(Integer)
-    releasedAt = Column(DateTime, index=True)
-    uuid = Column(String(36), primary_key=True)
-    advisorOrganizationUuid = Column(
-        ForeignKey('organizations.uuid'), index=True)
-    approvalStepUuid = Column(ForeignKey('approvalSteps.uuid'), index=True)
-    locationUuid = Column(ForeignKey('locations.uuid'), index=True)
-    principalOrganizationUuid = Column(
-        ForeignKey('organizations.uuid'), index=True)
-    legacyId = Column(Integer)
-    duration = Column(Integer)
-    customFields = Column(Text)
-
-    organization = relationship(
-        'Organization', primaryjoin='Report.advisorOrganizationUuid == Organization.uuid')
-    approvalStep = relationship('ApprovalStep')
-    location = relationship('Location')
-    organization1 = relationship(
-        'Organization', primaryjoin='Report.principalOrganizationUuid == Organization.uuid')
-    people = relationship("ReportPeople", back_populates="report")
-
-
-class ReportPeople(anet_logic_mixin):
-    __tablename__ = "reportPeople"
-
-    isPrimary = Column('isPrimary', Boolean, server_default=text("false"))
-    personUuid = Column('personUuid', ForeignKey('people.uuid'), index=True)
-    reportUuid = Column('reportUuid', ForeignKey('reports.uuid'), index=True)
-    isAttendee = Column('isAttendee', Boolean, server_default=text("true"))
-    isAuthor = Column('isAuthor', Boolean, server_default=text("false"))
-
-    __mapper_args__ = { 
-        "primary_key": [personUuid, reportUuid]
-    }
-
-    person = relationship("Person", back_populates="reports")
-    report = relationship("Report", back_populates="people")
